backend/requirements.txt
django
djangorestframework
numpy
pandas
tensorflow
pillow
scikit-learn
django-multiselectfield2




